Add script which calls factory-reset on a remote device
When a new image is deployed on a RevPi through LAVA, the device needs a
factory-reset with serial number, device type and mac address. As the
revpi might not be available when this script is started from LAVA, we
wait for the device until it is discoverable on the network. One could
surely make this with an oneliner in the LAVA user_commands, but this
solution is more readable and expendable.

To work correctly, the script needs four parameters, first the hostname
of the device, then the device type, then a serial number and the last
parameter is the mac address for the RevPi.

Signed-off-by: Frank Erdrich <f.erdrich@kunbus.com>