Copy Worker SSH-Key for root user
In order to test a RevPi image with our TestRack, the SSH-Key of the
LAVA-worker has to be available inside the image to be flashed.
This is why until now, RevPi images have always been modified before
being flashed on a RevPi device.
This means that the tested images are not the images that the customer
can download from our website, as these are altered.

With this commit this is modified to be able to use the RevPi images
that our customers have access to.

For this, after flashing a RevPi device, access it over ssh using the
default username and password and saving the Public Key of the worker
for pi and root users.

This way, the images can be tested without being previously altered and
the tests can be performed as before.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>