fix: Fix ability to flash multiple devices
So far, it was not possible to flash multiple devices with the ds3484
card, which is the one used for flashing devices automatically.

While on the one hand the hardware was practically ready, the software
structure still needed several adjustments.
The main adjustments were made in the device dictionary of the devices,
which is not documented here. These changes were made in the
lava-server, in the directory:
"/etc/lava-server/dispatcher-config/devices/".

On the other hand, after preparing the hardware and the structure
properly, it could be observed that the system still behaved
inadequately when flashing a device. This was due to the fact that the
Relay of the ds3484 board used to supply the voltage of the USB port of
a second device to be flashed, was not correctly deactivated.

The correct behavior of the board is achieved with the correction of this commit.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>