feat: Add configuration RevPi Connect/SE as Worker
So far there is only one worker in the TestRack, in which various LAVA
jobs are running. The worker is connected to the relay board ds3484,
which has 4 relays. 2 relays are used to control the voltage of DUTs
and 2 relays are used to control the voltage of USB. This would allow a
maximum of 2 RevPi devices to be flashed and tested.

To optimize this, a worker should be used for each RevPi configuration
that exists in the TestRack.
By using a RevPi Connect or RevPi Connect SE as a worker, it is
possible to flash the device and the relay output can also be used to
control the input voltage from the DUT.
The USB voltage is controlled with uhubctl.

This means that the relay board is no longer necessary.

This commit adds the script needed to program devices and the Pictory
configuration for RevPi Connect and RevPi Connect SE devices.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>