feat: Test rs485 - Worker as remote station
Enhance the Worker functionality by incorporating a new script to
enable a LAVA-Worker Connect to be able to operate as Remote Station
during rs485 test execution.

The script is prepared to work with the test from the
LAVA-test-definitions repository located here:
https://gitlab.com/revolutionpi/infrastructure/testing/
LAVA-test-definitions/automated/revpi/rs485/rs485.sh
The `rs485.sh`script is designed to wait for the execution of the
rs485-tx test on the DUT and subsequently run the rs485-rx test.

Also, the `worker_test_rs485.service` file is included to optimize the
execution of the script as a systemd daemon.

In the future the hardware for this test should be improved, as it is
not optimal to use another RevPi as a remote station for this kind of
test.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>