feat: add timeout loop for waiting for USB device
Added a timeout loop to improve the handling of waiting for the mass
storage device to appear after executing rpiboot.
The loop will check periodically until the device is found or the
maximum wait time is reached.
This prevents the script from waiting indefinitely and provides clearer
feedback when the device is not found within the specified time.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>