fix(factory-reset): wait to initiate reboot
When a DUT is flashed, it is booted and the script
revpi_factory_reset.sh is run on it. After the revpi-factory-reset on
the DUT was run successfully, the DUT is told to reboot. In order to
give the DUT some time to boot there is often a step in the job called
"wait_for_ssh" that waits until an SSH connection can be established
with the DUT, effectively waiting for it to reboot and be reachable
before continuing the job.

It's possible that, after revpi-factory-reset ran, the device doesn't
immediately reboot, but the job script that runs revpi-factory-reset on
the DUT does. This immediately starts the step in the job that waits for
an SSH connection to be established before continuing. If the check for
an SSH connection happens before the device reboots, this step is
immediately skipped and the worker doesn't wait for the device reboot.

Add a sleep at the end of the revpi_factory_reset.sh script of the
worker that gives the DUT a bit of time to reboot before exiting.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>