fix(factory_reset): don't fail on immediate reboot
Commit 2b57ca7 ("fix(revpi_factory_reset): abort on errors") adds "set
-e" to revpi_factory_reset.sh. This has the side effect of aborting the
script if a reboot of the DUT is immediately done instead of taking a
few seconds. In this case, ssh return 255 as the exit code, but the
behaviour is correct.

Because of this, handle the exit code of calling reboot with ssh on the
DUT manually, and don't let the script fail if 255 is returned.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>