feat(wait_for_ssh): show ssh output
The wait_for_ssh util fails to connect to the DUT fairly often, but it's
not easy to debug as ssh doesn't show any output, neither normal nor
errors.

Let ssh print output of any kind while trying to connect to the DUT.
Also increase the wait between tries to 10 seconds in order to not spam
the log with those messages.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>